+++ torchmetrics/classification/stat_scores.py
@@ -15,6 +15,7 @@ from typing import Any, Callable, Optional, Tuple
 
 import numpy as np
 import torch
+from deprecate import deprecated
 from torch import Tensor, tensor
 
 from torchmetrics.functional.classification.stat_scores import _stat_scores_compute, _stat_scores_update
@@ -132,6 +133,7 @@ class StatScores(Metric):
 
     """
 
+    @deprecated(target=True, deprecated_in="0.3.0", remove_in="0.4.0", args_mapping={"is_multiclass": "multiclass"})
     def __init__(
         self,
         threshold: float = 0.5,
@@ -145,6 +147,7 @@ class StatScores(Metric):
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
         dist_sync_fn: Callable = None,
+        is_multiclass: Optional[bool] = None,  # todo: deprecated, remove in v0.4
     ):
         super().__init__(
             compute_on_step=compute_on_step,